+++ data_viewer/plugins/builtin/plots/__init__.py
@@ -1,0 +1,2 @@
+"""Built-in declarative plot plugins."""
+